Add unit tests for LibraryContainer class methods
